split events.py into hydro.py and samples.py
