+++ wqio/core/__init__.py
@@ -6,9 +6,12 @@ from .features import (
     DataCollection
 )
 
-from .events import (
+from .samples import (
     CompositeSample,
     GrabSample,
+)
+
+from . hydro import (
     HydroRecord,
     Storm,
 )
