+++ wqio/tests/core_tests/baseline_images/core_tests/hydro_tests/test_stormplot.png
