+++ wqio/tests/core_tests/baseline_images/core_tests/samples_tests/test_Comp_without_storm_rug.png
