+++ wqio/tests/core_tests/baseline_images/core_tests/samples_tests/test_Grab_without_storm.png
